+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- don't trust session.portalHostname :(
+
 ## 0.20.3
 ### Fixed
 - GetBulk works with more than one group
+
 
 ## 0.20.2
 ### Fixed